cleaned up some code and bugs
