+++ src/components/LessonPage/Content.js
@@ -3,55 +3,31 @@ import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import styles from './Content.scss';
-import {processContent} from '../../utils/processContent';
+import {processContent, renderSnippets, createImage, createIframe} from '../../utils/processContent';
 import {getLessonContent} from '../../resources/lessonContent';
 import {getLessonPath} from '../../resources/lessonFrontmatter';
 import {setCheckboxesInDoc} from '../../utils/checkboxUtils';
 import {setCheckbox, removeCheckbox} from '../../reducers/checkboxes';
 import {getCheckboxesForLesson} from '../../selectors/checkboxes';
 
-const createIframe = () => {
-  const f = document.createElement('iframe');
-  f.id = 'makecoderenderer';
-  f.style.position = 'absolute';
-  f.style.left = 0;
-  f.style.bottom = 0;
-  f.style.width = '1px';
-  f.style.height = '1px';            
-  f.src = 'https://makecode.microbit.org/--docs?render=1';
-  document.body.appendChild(f);
-};
-
-const createImage = msg => {
-  let img = document.createElement('img');
-  img.src = msg.uri;
-  img.width = msg.width;
-  img.height = msg.height;
-  img.style.display = 'block';
-  img.style.margin = '0 auto 15px';
-  let code = document.getElementsByTagName('pre')[0];
-  if (code.className === 'microbit') {
-    code.parentElement.insertBefore(img, code);
-    code.parentElement.removeChild(code);
-  }
-};
-
-const renderSnippets = () => {
-  Array.from(document.getElementsByTagName('pre')).forEach(pre => {
-    const f = document.getElementById('makecoderenderer');
-    f.contentWindow.postMessage({
-      type: 'renderblocks',
-      id: pre.id,
-      code: pre.innerText
-    }, 'https://makecode.microbit.org/');
-  });
-};
-
 class Content extends React.PureComponent {
   createMarkup = () => {
     const {course, lesson, language, isReadme, isHydrated} = this.props;
     const lessonContent = getLessonContent(course, lesson, language, isReadme);
     return {__html: processContent(lessonContent, styles, isHydrated)};
+  };
+
+  createMicrobitImage = e => {
+    let msg = e.data;
+    if (msg.source != 'makecode') return;
+    switch (msg.type) {
+      case 'renderready':
+        renderSnippets();
+        break;
+      case 'renderblocks':
+        createImage(msg);
+        break;
+    }
   };
 
   updateCheckboxes = () => {
@@ -61,25 +37,23 @@ class Content extends React.PureComponent {
   };
 
   componentDidMount() {
-    createIframe();
-    window.addEventListener('message', ev => {
-      let msg = ev.data;
-      if (msg.source != 'makecode') return;
-      switch (msg.type) {
-        case 'renderready':
-          renderSnippets();
-          break;
-        case 'renderblocks':
-          createImage(msg);
-          break;
-      }
-    }, false);
     if (this.props.isHydrated) { this.updateCheckboxes(); } // When clicking in from different page
+    if (this.props.course === 'microbit') {
+      createIframe();
+      window.addEventListener('message', this.createMicrobitImage);
+    }
   }
 
   componentDidUpdate(prevProps) {
     const wasHydratedThisUpdate = !prevProps.isHydrated && this.props.isHydrated;
     if (wasHydratedThisUpdate) { this.updateCheckboxes(); } // When reloading page
+  }
+
+  componentWillUnmount() {
+    if (this.props.course === 'microbit') {
+      document.getElementById('makecoderenderer').remove();
+      window.removeEventListener('message', this.createMicrobitImage);
+    }
   }
 
   render() {
